Load theme after loading all libraries
By not loading the theme at the end of lib/appearance.zsh, but rather
after loading all libraries, a theme can rely on/override not just the basic
appearance stuff like PS1.
For example the 'dieter' theme intercepts (but does not alter) the accept-line
widget to change the display on the relevant keystroke.
Features like these are impossible or subject to race conditions without
this rearrangement.