Removing hard coded parens from rvm_prompt_info
I wanted to make use of rvm_prompt_info's error checking, but the
existin funciton had hard coded parens wrapping the ruby version. This
made for some odd behavior when using the system ruby (i.e. empty parens
in the prompt), and was not customizable. This should resolve the issue.

Updated risto theme to match.