Created so8res theme
This is a multi-line prompt theme. It's based off of Steve Losh's
[Extravagent Zsh](http://stevelosh.com/blog/2010/02/my-extravagant-zsh-prompt)
prompt. It has the usual cool git features (and can be used with
the untracked pull request that I made), but the coolest feature is
probably the battery life meter.

I use `acpi -b` to find battery info and display information accordingly.
You may need to modify the battery_* functions on your own to get it to
work on your machine, but it should work in the general case.