lib/git.zsh: Avoid calling current_branch()
The original code hard-coded 'origin' as the remote and called
current_branch() to get the corresponding remote branch name,
which may not necessarily be the correct upstream branch.

Use the the built-in @{upstream} syntax instead of calling
current_branch().

This improves performance and fixes a bug when the local and
remote branches do not have the same name.