Use git option zsh.ignore to disable prompt info
In the cases of very large repositories, where `git status` and friends
take a long time to complete, there should be a way to disable the smart
prompt info for a single repository, without needing to turn it off for
all of zsh.

This patch adds a call to `git config` at the beginning of the git
prompt subroutines to disable prompt info when the git option zsh.ignore
contains a non-empty value.

To disable the prompt info for a repository, use `git config zsh.ignore
yes` (or substitute your preferred value for "yes").  To re-enable the
info for a repository, use `git config --unset zsh.ignore`.