Being more accurate and careful with .zshrc during install/uninstall
install.sh:
  In the case where copy fails, remove will probably succeed.
For example, if the file system were full.  So we're trying
to be more careful with .zshrc here.

uninstall.sh:
  The messages are not accurate to what the script is doing
with .zshrc -- It says, "Backing up to .zshrc.pre-oh-my-zsh"
but the next line removes it.  So we should rename .zshrc
instead of removing it before we restore the original.