Debian plugin now pipes which stderr to stdout
The debian plugin calls which to determine whether aptitude & sudo are
installed.  However, if these are not installed then the error output
from which will be output.  Since this isn't really an error condition,
this commit updates the which calls to pipe stderr to stdout, thus
hiding the error messages.