Setup special keybindings using terminfo
The support for special keys right now is simply with hardcoded values,
this commit changes that to using terminfo information for the keys to
set the binds.  Should be much more general, and just work, at least
most of the time.  This also fixes a bug in the vi-mode plugin where the
special keys were not bound because setting vi mode removes binds set in
emacs mode.