Sublime plugin should check more locations
On OSX, the sublime plugin checks for various possible locations of the
Sublime Text application; on Linux, however, only two spots are covered.
This change uses the same detection mechanism used for OSX for Linux
systems to check for the typical install locations:

    $HOME/bin/sublime_text
    /usr/local/bin/sublime_text
    /usr/bin/sublime_text

This allows non-root users on Linux systems to use the plugin without
hassle.