fix virtualenv's with virtualenv plugin / prompt display
Virtual Env modifies the prompt in such a way that when used with
agnoster the virtualenv will be displayed twice.

Using the virtualenv oh-my-zsh plugin disables this behavior (simply
because it disables the virtual env prompt mangling) however when using
the plugin it would be nice to keep the single prompt display as
expected, and still allow for the function to provide prompt info which
other themes may use.