Fixed version of macports.plugin.zsh to fix issue #1983
As macports user i think it may be useful to put the export command
from macports in the plugin.
MacPorts adds this line when finishes its installation.
So if macports already installed, when oh-my-zsh install script runs,
the line disappear.
If macports is installed after oh-my-zsh, the added lines at the end of
the .zshrc need to be removed.

This update fix issue #1983.