Fix ssh-agent plugin on OS X.
In b2376c3, the `ps` command was changed away from `ps -ef` to
`ps x` (well, `ps -x`, but it was supposed to be `ps x`) because
OpenBSD doesn't support `-ef`. Sadly, OS X doesn't support `ps x`.
This fix adds a switch to use the old flags on OS X, and the newer,
more concise flags on other platforms.