prevent preventing headaches to prevent headaches
There are some verbose aliases for rm et al that are supposed to
"prevent headaches". However, the real headache is when you recursively
rm a directory with a million files in it and have to sit through all of
that output. Fix this by sticking to the tried-and-true default of
non-verbose mode. Users can always ask for verbose output when they want
it.