Close issues that are no longer needed
This is an empty commit to close issues and pull requests that were
opened some time ago and which the evolution of the repository have
made obsolete. The reason for closing is written next to the issue
to be closed.

Refer to the pull request to see the issues closed with this commit:
https://github.com/robbyrussell/oh-my-zsh/pull/2809