Fix to make the script works with any language
Hi,

The script fail when the language of the system is not set at "en_US.UTF-8" . I have made change to force the lang to en_US.UTF-8 when this script run.

I also notice the behead_re and  diverge_re regexp are wrong. (extra # and space at the begining)

Now the gitstatus.py return the correct informations.