plugins/themes: Rewrite theme() and lstheme() to remove pwd and variable side effects and make behavior consistent.
Include custom themes in lstheme() output.
Add comments that lay down the function interface.
Look for custom themes in $ZSH_CUSTOM/themes as well as base $ZSH_CUSTOM.
Rewrite theme's random theme selection in terms of lstheme() so random themes are included, and to streamline code.
Add a failure indication to theme() with nonzero return status. When a bad theme name is requested, indicate so by issuing a warning and returning nonzero.