Allow the systemd plugin to work when sudo is not installed.
Problem: The systemd plugin requires sudo to be installed.
The plugin set-up aliases sc-* for each systemctl command, and
prefix the alias with sudo so they can be used from non-root
account. However, this doesn't work if "sudo" is not installed.

Solution: Do not prefix the alias with "sudo" when they
are created for the root account.

This doesn't change anything for non-root user, but allows
root user to use the plugin even when sudo is not available
on the machine.