Fix incorrect PATH setting with system-wide rbenv
When using system-wide rbenv (as in provided by vendor packages like in Ubuntu or
FreeBSD), PATH is not set to the correct ruby binary.
This leads to being unable to use rubies installed with rbenv (and stick with the
system-wide ruby installed, if any).
This patch correct this behaviour, setting the correct PATH.