[hg] Allow turning off the `hg status` call in a similar fashion to the git plugin
In some very large repositories, `hg status` can take a pretty long time. It doesn't feel great to have a multi second delay every time you run a command while you wait for hg status to complete.

Mirroring the functionality in the git plugin, this diff adds a check for a new configuration option in `hg config`. If this option is enabled we disable the `hg status` check and just show the clean marker.