[hg] Cache the result of the `in_hg` call within `hg_prompt_info` helpers
The helper functions are smart only work if the user is in an hg repository. Unfortunately checking that fact can take a while for large repos, so these redundant checks are taking a while. This diff stores the result of the first call from `hg_prompt_info` in a local variable and just checks that for every other function.

This gave me a multi-second speedup locally.