github plugin: remove hub
It should not be included in the github plugin, because it is a tool by itself
and can cause some issues.

Apart from that it does (up to) 2 unnecessary calls to the `hub`
command.

After all, the integration boils down to:

    alias git=hub

This is not enough for an own plugin, but users should just add this to
their zshrc.