Deprecating brew cask plugin
I'm suggesting that this be deprecated in favor of the modified script in the
zsh-completions project. See https://github.com/joshka/zsh-completions/commit/31b4e82887626eee236a9806ec9c4f5dee90056c

A copy of the relevant part of the commit message of the above is included here:

The original plugin suffers from the following issues (IMO):
1. it is part of oh-my-zsh which some have chosen not to use
2. it acts as a script overriding the existing brew completion rather than as a
standard autoloadable file, this means it must be manually installed rather than
automatically installed as part of compinit.

The changes from the plugin are to fix item 2, as well as update the commands to
those current in homebrew cask. This change relies on a change pending in the
homebrew completion to allow it to call subcommand completions (see
https://github.com/joshka/brew/commit/6b010c5700095685bad6797038e83ef558df5104)
and hence should not be merged until that has been merged.

Ideally, this completion would be installed as part of the homebrew cask
installation, but there doesn't seem to be an easy way to do so as it is
homebrew cask is installed as a homebrew tap. It also doesn't really fit in the
homebrew main repo for similar reasons.