Tag support for detached branche in git_prompt_info
If the HEAD is not a symbolic ref (e.g. refs/heads/master) but it points
to a commit ID (branch is detached), the git_prompt_info shows commit ID
even if the commit ID points has a tag. This patch allows to display the
tag version if there is a tag for the HEAD commit ID.