Remove misleading comment about `export`
`export` is not optional unless the parameters are already part of the environment, either due to being exported beforehand or by virtue of already being in the environment when `zsh` started (e.g. `HOME` or `PATH`). Although the description (intentionally?) reads as if `dotenv` is specifically made to only set environment variables, it does not actually do anything to automatically export parameters into the environment. Stating that `export` is optional may lead to confusion, for example this question on StackOverflow: https://stackoverflow.com/q/43878291/2992551