Adding plugin node-modules-path
If chpwd is changed to a path where any folder within that subpath
contains a node_modules folder, then that path + .bin is added to the
PATH env variable.

The reason for writing this plugin is that I often open up atom directly
from the shell in a directory of interest. If I have flow installed
locally then I have to add the node_modules folder in that path to PATH
before I open up Atom.