Create Fasd cache file in user home directory
If oh-my-zsh is installed globally, ZSH_CACHE_DIR  may point to a directory which is not writable by the user or which is shared amog all the users. For example, in Arch Linux, it is pointing to /usr/share/oh-my-zsh.

This fix will create Fasd cache file as hidden in his home directory.