Set RBENV_ROOT to "$HOME/.rbenv" if not already set
This is the default behavior of rbenv and what users are expecting
most of the time.

It allows users to have their own set of rubies and gems. It also
prevents losing all rubies when rbenv is updated using Homebrew which
is not true when RBENV_ROOT is set to /usr/local/opt/rbenv.