Unbreak vi's visual mode
Vi (and zsh's vi mode) a few modes, one of them being visual mode, which
is triggered by pressing "v".

Entering it using "v" is the default on vi/vim/nvim, and also on zsh
when using `bindkey -v`.

However, the vi-mode plugin overrides this with a custom action,
effectively making this mode inaccessible and breaking vi mode.

This PR undoes that, since it seems to be a huge oversight when mapping
that particular key, and makes vi's visual mode accessible when using the
vi-mode plugin.