Add SHOW_DIRTY_BRANCH option
This option is enabled by default (so that there's no regression). It
simply allows to enable/disable the dirty text in the prompt.

This is because in huge repositories, like the one I work with, asking
for the status can take several hundreds of milliseconds.

This is a not so intrusive way of providing an option in this case.