Adds Homebrew and completion support to asdf plugin
* Modifies the search logic for asdf to include Homebrew when it is
installed.  The implementation is adapted from the pyenv plugin.  Per
feedback from @mcornella, asdf is either going to be in the
Homebrew directory or $HOME/.asdf.  Since no distributions appear
to package asdf, the other locations are unncecessary.  This new
assumption allowed the search logic to be simplified as well.

* Includes command completions for both regular and Homebrew installs.
This enhancement was inspired by PR $6045.