agnoster: speedup for repos with enabled hide-dirty
For huge repos, status checking can take many seconds. It's pretty annoying to
have such delay on each prompt message inside such repo. agnoster uses
parse_git_drity which respect the oh-my-zsh.hide-dirty flag. Thus, we don't
perform an expensive check in this case. However, it doesn't solve the original
problem because we still have a delay because of check-for-changes. The line
should be executed only for dirty repos.