Allow skipping execution of compinit/compaudit
Loading autocompletions and auditing the directories for insecurities is
a relatively time-consuming process in the lifecycle of a shell process
startup. If a user would like to skip this process (either one-off or on
a regular basis), they may do so after this change by defining
`ZSH_DISABLE_COMPINIT`.

The change permits the variable to be defined in any way and merely
checks that it exists, not what its value is, using a Zsh idiom
involving treating the variable as a substitution within an arithmetic
expression.

After changing this, it will be necessary to add information about this
change here: https://github.com/robbyrussell/oh-my-zsh/wiki/Design.

Fixes #7336.