change AWS_DEFAULT_PROFILE to AWS_PROFILE
The environment variable name used to be AWS_DEFAULT_PROFILE but the
CLI documentation now only mentions AWS_PROFILE.

https://docs.aws.amazon.com/cli/latest/userguide/cli-environment.html

It seems like the CLI was the only tool using AWS_DEFAULT_PROFILE, and
all the AWS SDKs used AWS_PROFILE, so they standardized on it.

https://onetechnical.wordpress.com/2016/10/07/the-curious-case-of-aws_default_profile/

Note: still left AWS_DEFAULT_PROFILE on the method to set the profile to
maintain backwards compatibility.