Add alias glogb='glog --branches' to git log only local branches
For a repo with a large number of remote branches, `gloga` is too noisy.
Replacing `--all` with `--branches` does the trick by only showing local
branches. Here we add a new alias which is the same as `gloga` except for the
`--all` argument, which is swapped out for `--branches` as mentioned earlier.