Clean up the general aliases in the kubectl plugin
There's a ton of aliases being added for kubectl commands, which is
awesome, but if you're like me you only remember a handufl of them.
Really the top level aliases are usually the most valuable for me and
easier to remember.

This change adds a few missing `general` aliases, and groups the general
aliases together in the top of the doc.  This way even if "your
specific" alias is missing, all the primitives are in place to provide
the foundation and make kubectl much easier to work with.

Added aliases:
  * kg    - kubectl get
  * kgans - kubectl get $@ --all-namespaces (kga is taken)
  * kd    - kubectl describe