plugins/vi-mode: Improve keymap tracking
Makes sure that `VI_KEYMAP` starts as `main`, and that it's reset on
`zle-line-init`. This is better than binding ^J and ^M (which bindings
have now been removed), because those may already be mapped to
something, *and* they are not the only key sequences can cause the line
to re-initialize (such as Ctrl-C).

The `zle-line-finish` is defined somewhat redundantly here, since it's
defined the same lin lib/key-bindings.zsh, but is used for cursor shape
control in a subsequent commit.