Add ZSH_TMUX_NEW_SESSION_ARGS option to tmux plugin
When the tmux plugin auto-starts tmux, it calls `tmux new-session`
with no arguments.  It can be useful to pass additional arguments
to this command to set the session name or session group.
This commit adds a new environment variable ZSH_TMUX_NEW_SESSION_ARGS
that can be used to pass additional arguments to the tmux new-session
command.