fix(pipenv): error when pipfile is a directory
This fixes a bug when `Pipfile` is a directory which cuases `pipenv` to
crash.

To reproduce it:

* create a directory `Pipfile`
* enter the directory

You are going to see an output similar to this:

```
Traceback (most recent call last):
  File "/Users/rmax/.pyenv/versions/3.8.2/bin/pipenv", line 10, in <module>
    sys.exit(cli())
  ...
  File "/Users/rmax/.pyenv/versions/3.8.2/lib/python3.8/site-packages/pipenv/project.py", line 677, in touch_pipfile
    with open("Pipfile", "a"):
IsADirectoryError: [Errno 21] Is a directory: 'Pipfile'
```

The current code does a `-a` check which is true for any file type,
including directories. A better way is to use `-f` to ensure `Pipfile`
is a regular file.