git: make git_main_branch configurable
Use `GIT_MAIN_BRANCH_PRIORITY`, which is a list of branch names
separated by ":", to make the function `git_main_branch` configurable.

Now `git_main_branch` checks for all branches in the list and returns
the first existing branch in the list.

The default `GIT_MAIN_BRANCH_PRIORITY` is `master:main` instead of
`main:master`, and `git_main_branch` returns `master` if no branch in
`GIT_MAIN_BRANCH_PRIORITY` exists, so that it's not a breaking change.