Add new theme based on agnoster.
This theme keeps the same ideals as agnoster's theme on only showing
relevant information.  The main addition is a right side prompt with
components similar to how the left side prompt is set up.

It also removes some symbols that messed up rendering on my chromebook
(because the terminal emulator for it sucks).