✨ Support using mvnw in multi-module projects
Frequently Maven projects are multi-module, so there's a top level
directory that has the mvnw script and a parent POM file and a number of
subdirectories that have the their own POM files but no mvnw wrapper.
Walk one directory up to see if there's a mvnw and execute it,
supporting this use case.

Fixes #8557