sudo: Support aliases to $EDITOR
If you use some alias for your $EDITOR, say `e`. You type `e /to/file`.
But that file needs sudo permission. You naturally do <Esc><Esc><Enter>.
But, to your dismay, it errors out `command not found`, because it
didn't detect your alias to $EDITOR, thereby failing to convert it to
`sudoedit`.

This commit tries to detect your alias and convert it to sudoedit.