fix(magic-enter): make it compatible w/ zsh-autosuggestions plugin
This solution makes the magic-enter function a wrapper for the accept-line
zle widget, meaning that magic-enter will be run instead of accept-line
(this is usually done when pressing Enter or running `zle accept-line`).

The zsh-autosuggestions plugin will automatically notice this is a wrapper
and properly assign it the "clear suggestions" action when run.

Also: refactored and changed code style of the plugin.