Update README.md for enabling nvm plugin settings
It wasn’t clear to me where and how I should set the `NVM_AUTOLOAD` variable, i.e. `NVM_AUTOLOAD=1` or `export NVM_AUTOLOAD=1`. It seems the clue was in one of the other values, but I skipped that as that wasn’t the setting I was looking to change. Hopefully this is a bit clearer.